+++ internal/db_migrations/migrations/000013_add_lautern_duration.down.sql
@@ -1,0 +1,2 @@
+ALTER TABLE "summaries"
+DROP COLUMN lautern_duration;